🚚 Rename files to plural
