+++ src/response/rich-texts.ts
